Vitaaaa !

+++ src/main/java/org/vehicule/wsvehicule/controller/business/KilometrageController.java
@@ -4,43 +4,62 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.vehicule.wsvehicule.model.business.Kilometrage;
 import org.vehicule.wsvehicule.service.KilometrageService;
+import org.vehicule.wsvehicule.service.TokenResponseService;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @RestController
 @RequestMapping("/kilometrages")
 public class KilometrageController {
-    KilometrageService kilometrageService;
+    private final KilometrageService kilometrageService;
+    private final TokenResponseService tokenResponseService;
 
     @Autowired
-    public KilometrageController(KilometrageService kilometrageService) {
+    public KilometrageController(KilometrageService kilometrageService, TokenResponseService tokenResponseService) {
         this.kilometrageService = kilometrageService;
+        this.tokenResponseService = tokenResponseService;
     }
 
     @GetMapping
-    public List<Kilometrage> readAll() {
-        return kilometrageService.readAll();
+    public List<Kilometrage> readAll(@RequestHeader Map<String,String> headers) {
+        if (tokenResponseService.validateAuthorization(headers))
+            return kilometrageService.readAll();
+        else
+            throw new RuntimeException("- Access Denied -");
     }
 
     @PostMapping
-    public void createKilometrage(@RequestBody Kilometrage kilometrage) {
-        kilometrageService.create(kilometrage);
+    public void createKilometrage(@RequestHeader Map<String,String> headers, @RequestBody Kilometrage kilometrage) {
+        if (tokenResponseService.validateAuthorization(headers))
+            kilometrageService.create(kilometrage);
+        else
+            throw new RuntimeException("- Access Denied -");
     }
 
     @GetMapping("/{id}")
-    public Optional<Kilometrage> readKilometrage(@PathVariable Long id) {
-        return kilometrageService.read(id);
+    public Optional<Kilometrage> readKilometrage(@RequestHeader Map<String,String> headers, @PathVariable Long id) {
+        if (tokenResponseService.validateAuthorization(headers))
+            return kilometrageService.read(id);
+        else
+            throw new RuntimeException("- Access Denied -");
     }
 
     @PutMapping("/{id}")
-    public void updateKilometrage(@PathVariable Long id, @RequestBody Kilometrage kilometrage) {
-        kilometrageService.update(id, kilometrage);
+    public void updateKilometrage(@RequestHeader Map<String,String> headers, @PathVariable Long id, @RequestBody Kilometrage kilometrage) {
+        if (tokenResponseService.validateAuthorization(headers))
+            kilometrageService.update(id, kilometrage);
+        else
+            throw new RuntimeException("- Access Denied -");
     }
 
     @DeleteMapping("/{id}")
-    public void deleteKilometrage(@PathVariable Integer id) {
-        kilometrageService.delete(id);
+    public void deleteKilometrage(@RequestHeader Map<String,String> headers, @PathVariable Integer id) {
+        if (tokenResponseService.validateAuthorization(headers))
+            kilometrageService.delete(id);
+        else
+            throw new RuntimeException("- Access Denied -");
     }
 
 }
